feat(user): Update check security

+++ src/main/java/mingzuozhibi/coreserver/modules/user/UserController.java
@@ -11,15 +11,14 @@ import mingzuozhibi.coreserver.modules.user.enums.Role;
 import mingzuozhibi.coreserver.security.auth.SessionManager;
 import mingzuozhibi.coreserver.security.support.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Objects;
-import java.util.Set;
 
 @RestController
+@PreAuthorize("hasAnyRole('RootAdmin','UserAdmin')")
 public class UserController extends BaseController {
 
     @Autowired
@@ -31,23 +30,20 @@ public class UserController extends BaseController {
     @MsgsWired(Index.User)
     private Msgs msgs;
 
-    @Transactional
+    @Transactional(readOnly = true)
     @GetMapping("/api/users")
-    @PreAuthorize("hasRole('UserAdmin')")
     public String findAll(@PageDefault(size = 20) PageParams params) {
         return objectResult(userRepository.findAll(params.toPageable()));
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     @GetMapping("/api/users/{id}")
-    @PreAuthorize("hasRole('UserAdmin')")
     public String findById(@PathVariable Long id) {
         return userRepository.findById(id, this::objectResult);
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     @GetMapping("/api/users/find/username/{username}")
-    @PreAuthorize("hasRole('UserAdmin')")
     public String findByUsername(@PathVariable String username) {
         return userRepository.findByUsername(username, this::objectResult);
     }
@@ -60,12 +56,12 @@ public class UserController extends BaseController {
 
     @Transactional
     @PutMapping("/api/users/{id}/enabled")
-    @PreAuthorize("hasRole('UserAdmin')")
     public String setEnabled(@PathVariable Long id, @RequestBody EditForm form) {
         return userRepository.findById(id, user -> {
-            user.getRoles().stream()
-                .filter(Set.of(Role.RootAdmin, Role.UserAdmin)::contains)
-                .forEach(this::checkSecurityOfRole);
+            checkRootAdminUser(user);
+            if (user.isRootAdmin()) {
+                throw new SecurityException("不能禁用RootAdmin用户");
+            }
             if (user.isEnabled() != form.enabled) {
                 user.setEnabled(form.enabled);
                 sessionManager.deleteSession(user);
@@ -78,10 +74,10 @@ public class UserController extends BaseController {
 
     @Transactional
     @PostMapping("/api/users/{id}/roles")
-    @PreAuthorize("hasRole('UserAdmin')")
     public String pushRole(@PathVariable Long id, @RequestBody EditForm form) {
         return userRepository.findById(id, user -> {
-            checkSecurityOfRole(form.targetRole);
+            checkRootAdminUser(user);
+            checkUserAndRole(user, form.targetRole);
             if (user.getRoles().add(form.targetRole)) {
                 sessionManager.deleteSession(user);
                 msgs.info("用户%s给%s添加了权限%s",
@@ -93,10 +89,10 @@ public class UserController extends BaseController {
 
     @Transactional
     @DeleteMapping("/api/users/{id}/roles")
-    @PreAuthorize("hasRole('UserAdmin')")
     public String dropRole(@PathVariable Long id, @RequestBody EditForm form) {
         return userRepository.findById(id, user -> {
-            checkSecurityOfRole(form.targetRole);
+            checkRootAdminUser(user);
+            checkUserAndRole(user, form.targetRole);
             if (user.getRoles().remove(form.targetRole)) {
                 sessionManager.deleteSession(user);
                 msgs.info("用户%s从%s移除了权限%s",
@@ -106,17 +102,24 @@ public class UserController extends BaseController {
         });
     }
 
-    private void checkSecurityOfRole(Role role) {
-        Objects.requireNonNull(role);
-        SecurityUtils.doSecurityCheck(roles -> {
-            if (Role.RootAdmin == role) {
-                return false;
-            }
-            if (Role.UserAdmin == role && !roles.contains(Role.RootAdmin)) {
-                return false;
-            }
-            return true;
-        });
+    private void checkRootAdminUser(User targetUser) {
+        if (!isSameUser(targetUser) && targetUser.isRootAdmin()) {
+            throw new SecurityException("不能变更RootAdmin用户");
+        }
+    }
+
+    private void checkUserAndRole(User targetUser, Role targetRole) {
+        if (targetRole == Role.RootAdmin) {
+            throw new SecurityException("不能变更RootAdmin权限");
+        }
+        if (isSameUser(targetUser) && targetRole == Role.UserAdmin) {
+            throw new SecurityException("不能删除自己的UserAdmin权限");
+        }
+    }
+
+    private boolean isSameUser(User targetUser) {
+        User currentUser = SecurityUtils.getUserOrThrow();
+        return Objects.equals(currentUser, targetUser);
     }
 
 }
